--- bambi/backends/__init__.py
@@ -1,7 +1,0 @@
-from .pymc3 import PyMC3BackEnd
-from .stan import StanBackEnd
-
-__all__ = [
-    'PyMC3BackEnd',
-    'StanBackEnd'
-]
